Update: Implemented first version of parameter estimate
Not yet working. Problem with extimating a/alpha coefficients.
What's the problem? Only more time needed?

Implemented vanilla version of calulating splines and the
corresponding derivatives to make it independent of b factors.
The benifit is that computations are model/data independent and
hence only need to be computed once.

Using difference to equilibrium instead of total values.