Bugfix: Fixed conversion method. Yet new bugs not yet fixed
Method finds minimum and converges. However, result doesn't
really make sense yet. Changes in restriction led to finding new bugs.
These were not yet fixed.

Good results are, tho, that the parameters for a/alpha are for all positions
comparatively close, suggesting that all positions are driven by the same
dynamics. 85% of the values are in a very short range of values. Yet
outliers can also lead to very large or small values.