Update: changed breaking condition
The condition to stop the optimisation procedure is
now defined via the ratio of the parameters that does not change
sufficiently anymore. A minimum threshold is set (ex 90 percent)
and then afterwards the procedure stops when the ratio does not
change significantly anymore. Alternatively, you can set a maximum
number of iterations.