Update: Removed
As only dummy data is uploaded to the git repository,
there's no reason why the main function with directory paths
to the data is added too. Thus, the original 1d react file was
split into test_main.py and main.py. While test_main.py is working
fine with the dummy data provided in the data directory in
the repository, the main.py was added to the gitignore file